requirements.txt
PyYAML
gns3-netifaces
gspread
oauth2client
PyOpenSSL
sshpubkeys
